requirements/base.txt
celery==4.2.1
certifi==2017.11.5
chardet==3.0.4
discogs-client==2.2.1
dj-database-url==0.5.0
dj-static==0.0.6
django-cacheback==1.4.0
django-cors-headers==2.1.0
django-environ==0.4.4
django-extensions==1.9.8
django-redis==4.9.1
djangorestframework-filters==1.0.0.dev0
django==2.0.8
djangorestframework==3.9.0
factory-boy==2.11.1
flake8==3.5.0
gevent==1.3.5
gunicorn==19.9.0
idna==2.6
lxml==4.1.1
markupsafe==1.0
oauthlib==2.0.6
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
requests==2.20.0
selenium==3.7.0
six==1.11.0
sqlparse==0.2.4
typing==3.6.2
urllib3==1.22
whitenoise==3.3.1
xmltodict==0.11.0